renamed Turnips to StalkMarket and TurnipsBuyDay to DayPrices moved StalkMarket Property to the Town class
