+++ MyHorizons/Data/Save/MainSaveFile.cs
@@ -3,7 +3,6 @@ using MyHorizons.Data.TownData;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using MyHorizons.Data.TurnipsData;
 
 namespace MyHorizons.Data.Save
 {
@@ -13,7 +12,6 @@ namespace MyHorizons.Data.Save
         
         public int NumPlayers => _playerSaves.Count;
         public readonly Town? Town;
-        public readonly Turnips? Turnips;
 
         public MainSaveFile(in string headerPath, in string filePath)
         {
@@ -29,7 +27,6 @@ namespace MyHorizons.Data.Save
                 }
 
                 Town = new Town(this);
-                Turnips = new Turnips(this);
             }
         }
